課題7完成

+++ app/controllers/favorites_controller.rb
@@ -16,9 +16,12 @@ class FavoritesController < ApplicationController
   end
   
   def destroy
-    favorite = Favorite.find_by(user_id: current_user.id, topic_id: params[:topic_id])
     
-    favorite.destroy
+    favorite = Favorite.find_by(params[:topic_id])
+    favorite.destroy if favorite
+    flash[:success] = "登録解除しました！"
+    redirect_back(fallback_location: root_path)
+    
   end 
   
 end
